adjusted handler to handle errors, created 2 pages, modified actions
